Move round manager into standalone module
