+++ README.md
@@ -190,10 +190,10 @@ Combat System
 Modular combat using a 2-second tick:
 
 ```python
-from combat.combat_manager import CombatRoundManager
+from combat.round_manager import CombatRoundManager
 CombatRoundManager.get().start_combat([fighter1, fighter2])
 ```
-The previous `combat.round_manager` module now simply re-exports these
+The old `combat.combat_manager` module now simply re-exports these
 classes and is considered deprecated.
 
 Supports action queues, status effects, resistances
