+++ typeclasses/tests/test_combat_full_fight.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from evennia.utils.test_resources import EvenniaTest
 from combat.combat_actions import Action, CombatResult
-from combat.combat_manager import CombatRoundManager, CombatInstance
+from combat.round_manager import CombatRoundManager, CombatInstance
 
 
 class DamageAction(Action):
